requirements.txt
mcp
openai
sqlalchemy
pymysql
pyjwt[crypto]
